+++ src/thirtytwo/Win32/Graphics/Gdi/HBRUSH.cs
@@ -24,7 +24,7 @@ public unsafe partial struct HBRUSH : IDisposable
 
     public void Dispose()
     {
-        if (Value != 0 && Value != -1)
+        if ((nint)Value != 0 && (nint)Value != -1)
         {
             Interop.DeleteObject(this);
             Unsafe.AsRef(in this) = default;
