+++ src/thirtytwo/Win32/Graphics/Gdi/HGDIOBJ.cs
@@ -7,9 +7,9 @@ public unsafe partial struct HGDIOBJ
 {
     public OBJ_TYPE GetObjectType() => (OBJ_TYPE)Interop.GetObjectType(this);
 
-    public static implicit operator HGDIOBJ(StockFont font) => new(Interop.GetStockObject((GET_STOCK_OBJECT_FLAGS)font));
-    public static implicit operator HGDIOBJ(StockBrush brush) => new(Interop.GetStockObject((GET_STOCK_OBJECT_FLAGS)brush));
-    public static implicit operator HGDIOBJ(SystemColor color) => new(Interop.GetStockObject((GET_STOCK_OBJECT_FLAGS)color));
-    public static implicit operator HGDIOBJ(StockPen pen) => new(Interop.GetStockObject((GET_STOCK_OBJECT_FLAGS)pen));
-    public static implicit operator HGDIOBJ(GET_STOCK_OBJECT_FLAGS stockObject) => new(Interop.GetStockObject(stockObject));
+    public static implicit operator HGDIOBJ(StockFont font) => new((nint)Interop.GetStockObject((GET_STOCK_OBJECT_FLAGS)font));
+    public static implicit operator HGDIOBJ(StockBrush brush) => new((nint)Interop.GetStockObject((GET_STOCK_OBJECT_FLAGS)brush));
+    public static implicit operator HGDIOBJ(SystemColor color) => new((nint)Interop.GetStockObject((GET_STOCK_OBJECT_FLAGS)color));
+    public static implicit operator HGDIOBJ(StockPen pen) => new((nint)Interop.GetStockObject((GET_STOCK_OBJECT_FLAGS)pen));
+    public static implicit operator HGDIOBJ(GET_STOCK_OBJECT_FLAGS stockObject) => new((nint)Interop.GetStockObject(stockObject));
 }
