+++ src/thirtytwo/Win32/Graphics/Gdi/HPEN.cs
@@ -21,7 +21,7 @@ public unsafe partial struct HPEN : IDisposable
 
     public void Dispose()
     {
-        if (Value != 0 && Value != -1)
+        if ((nint)Value != 0 && (nint)Value != -1)
         {
             Interop.DeleteObject(this);
             Unsafe.AsRef(in this) = default;
